feat(constants): move css-breakpoints constant
