+++ projects/evo-ui-kit/src/public_api.ts
@@ -64,6 +64,7 @@ export * from './lib/components/evo-portal/index';
 
 export * from './lib/common/evo-control-state-manager/evo-control-state-manager';
 export * from './lib/common/form-helper/form-helper';
-export * from './lib/common/constants';
+export * from './lib/common/constants/css-breakpoints';
+export * from './lib/common/constants/evo-color-palette';
 export * from './lib/common/evo-base-control';
 export * from './lib/common/Serializable';
